Adapt to geogig core module package refactoring to allow jar signing across modules
