+++ src/community/geogig/src/main/java/org/geogig/geoserver/rest/RepositoryResource.java
@@ -14,8 +14,8 @@ import java.util.Map;
 import org.geogig.geoserver.config.RepositoryInfo;
 import org.geoserver.rest.PageInfo;
 import org.geoserver.rest.format.FreemarkerFormat;
-import org.locationtech.geogig.rest.StreamingWriterRepresentation;
 import org.locationtech.geogig.rest.RestletException;
+import org.locationtech.geogig.rest.StreamingWriterRepresentation;
 import org.locationtech.geogig.rest.Variants;
 import org.locationtech.geogig.rest.repository.DeleteRepository;
 import org.locationtech.geogig.web.api.StreamWriterException;
