+++ src/community/geogig/src/test/java/org/geogig/geoserver/web/data/store/geogig/GeoGigDataStoreEditPanelTest.java
@@ -9,7 +9,7 @@ import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
 import static org.junit.Assert.assertNull;
 import static org.junit.Assert.assertTrue;
-import static org.locationtech.geogig.model.RevObjectTestSupport.hashString;
+import static org.locationtech.geogig.model.impl.RevObjectTestSupport.hashString;
 import static org.mockito.Matchers.anyString;
 import static org.mockito.Mockito.mock;
 import static org.mockito.Mockito.when;
